Fixes #23017 - Remove remote_task only if it exists

+++ app/lib/actions/proxy_action.rb
@@ -176,7 +176,7 @@ module Actions
     end
 
     def clean_remote_task(*_args)
-      remote_task.destroy!
+      remote_task.destroy! if remote_task
     end
 
     private
